requirements.txt
alembic==0.8.4
bleach==1.4.2
blinker==1.4
cffi==1.5.2
cssmin==0.2.0
Flask==0.10.1
Flask-Assets==0.11
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-SocketIO==2.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
gunicorn==19.4.5
hiredis==0.2.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.8
jsmin==2.2.0
Mako==1.0.3
MarkupSafe==0.23
misaka==2.0.0
Pillow==3.1.1
psycopg2==2.6.1
pycparser==2.14
python-editor==0.5
python-engineio==0.8.8
python-socketio==1.0
redis==2.10.5
six==1.10.0
SQLAlchemy==1.0.12
webassets==0.11.1
Werkzeug==0.11.4
WTForms==2.1
